Changed float type to double in rotation field.

+++ src/cls/setup/light.hpp
@@ -45,8 +45,10 @@ namespace arc
             //  -- Transformations --
             const math::Vec<3> m_trans; //! Translation to apply to the mesh.
             const math::Vec<3> m_dir;   //! Direction to face the mesh.
-            const float        m_rot;   //! Rotation to apply to the mesh.
+            const double       m_rot;   //! Rotation to apply to the mesh.
             const math::Vec<3> m_scale; //! Scalting to apply to the mesh.
+
+            //  -- Properties --
 
 
             //  == INSTANTIATION ==
